PRR-45 [chore]: 개인 정보 제공 관련 컴포넌트 이름 변경
