--- src/components/domain/MyPage/Profile/BasicInformation/index.tsx
@@ -1,3 +1,0 @@
-import BasicInformation from './BasicInformation';
-
-export default BasicInformation;
